LLMeetLivekitBackend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
bcrypt==4.0.1
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
livekit==1.0.8
livekit-api==1.0.2
livekit-protocol==1.0.3
loguru==0.7.3
multidict==6.4.4
mysql-connector-python==9.3.0
numpy==2.2.6
opencv-python==4.11.0.86
passlib==1.7.4
propcache==0.3.1
protobuf==6.31.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.0
